CRA-13: Changed the behaviour of the "use-for" keyword under well, so that wells are excluded from estimation due to synthetic logs or deviation only if these keywrods are not used. When they are used, they override deviation and synthetic well log tests.
Included a new enum in modelsettings, indicator, which is NO=0, YES=1 or NOTSET=2. The intention is that NOTSET by default behaves like YES, but we can distinguish it from a set YES.

The class variables useFor... in welldata.h has been changed to int, to use the enum above.

In welldata.cpp, the routines lookForSyntheticVsLog and calculateDeviation now only override the useFor... variables if these are NOTSET.

In xmlmodelfile.cpp, the reading of the <use-for...> keywords under <well> has changed to match the new setting with three possbilities.


git-svn-id: https://code.nr.no/scm/svn/crava/trunk@373 176de4d3-4747-0410-aa96-47f6d6634214