CRA-69 Generated facies probability quality indicator using chi square test.
faciesprob.cpp, h: Made functions calculateChiSquareTest and writeBWFaciesProb. In calculateChiSquareTest, a goodness of fit test for the wells is generated using chi square statistics. The class nrlib/random/chisquared.hpp is included to calculate the p-values of the test. The alternative goodness of fit test calculateChiSquareAlternativeTest is made, but is not called from crava. 

In writeBWFaciesProb, the facies probabilities are given to blockedlogs.cpp. 

In crava.cpp, the functions calculateChiSquareTest and writeBWFaciesProb are called. 

In Makefile, Fullmodules: random is included to be able to use chisquared. 

The blocked log in test case 9 is updated to include facies probabilities in the output file.

In blockedlogs.h, function getDz is made.

Have added some files in crava; hence crava.vcproj is updated.

git-svn-id: https://code.nr.no/scm/svn/crava/trunk@484 176de4d3-4747-0410-aa96-47f6d6634214