CRA-87 Added tasklist.cpp, h.
tasklist.cpp is a class containing a list of suggested tasks to be done for the user. Each time CRAVA gives a warning, a new task should be added using TaskList::addTask(""). The task contains a suggestion about how to solve the problem that gave the warning.

The task list is written to the log-file from main.cpp.
The class tasklist.cpp is included in crava.vcproj

git-svn-id: https://code.nr.no/scm/svn/crava/trunk@501 176de4d3-4747-0410-aa96-47f6d6634214