CRA-102 Fixed bug such than CRAVA terminates if the facies names in the xml-file are not identical to the facies names in the wells. Have also made facies estimation more consistent by writing information to log-file and giving tasks to taskList when CRAVA gives out warnings.
xmlmodelfile.cpp: Included check to see that facies probabilities read from file are in interval [0,1]. Crava terminates if this is not true.

model.cpp: 
Included tasklist.h. 
Added the number of facies to nGrids in checkAvailableMemory() for the case when the prior facies probabilities are given in probability cubes. 
printSettings(): write out prior facies names and probabilities to header of log-file when these are given in the xml-file. 
readStormFile(): Fixed bug making padding for prior facies probabilities from storm file even when padding is not requested.
processWells(): Added tasks to the tasklist when warnings are given. Added error message when this is given in checkFaciesNames. 
checkFaciesNames(): Included inputFiles. Compare names in xml-file to names in wells. Give error when these are not identical. 
processPriorFaciesProb(): Added tasks when warnings are given. Write facies and corresponding probabilities to log-file under "Process facies probabilities", when these are given in the xml-file.

model.h: Included inputfiles in checkFaciesNames(). 

faciesprob.cpp: Included tasklist.h. Added information in normalizeCubes() to be able to give correct tasklist when the probability cubes in the xml-file are not in interval [0,1] and do not sum to one. Added exception in calculateChiSquareTest() when degrees of freedom is zero due to only one actual facies class. 

fftgrid.h: Corrected comments explaining nx_, ny_, nz_, nxp_, nyp_, nzp_.

referencemanual.tex: Changed <probabilitycube> to <probability-cube> to obtain consistency with CRAVA.  

git-svn-id: https://code.nr.no/scm/svn/crava/trunk@502 176de4d3-4747-0410-aa96-47f6d6634214