CRA-135 Removed several bugs in xmlmodelfile.cpp making sure the correct grid-outputs are written to file in Model::printSettings().
Made elasticOutput_ in modelsettings.cpp, being true if elastic output is written to file. 

Made simbox::setTopBotName(). The names were previously set in simbox::writeTopBotGrids(), giving error if this function was not called.

Set log-level to high in all test cases in the test_suite to make sure the correct information is written to the log-file. With the level being low, it is impossible to know exactly what is written to the log-file.
 

git-svn-id: https://code.nr.no/scm/svn/crava/trunk@585 176de4d3-4747-0410-aa96-47f6d6634214