CRA-231 Prepared model file for 4D. <survey> is now repeated for each time lapse, and the keyword <vintage> is added under <survey>. <vintage> contains <year>, <month> and <day-of-month>.
All variables in ModelSettings and InputFiles that are set under <survey> are updated to contain time lapse data. 

main.cpp is reconstructed such that it loops over all time lapses. modelavodynamic.cpp is called for each time lapse, and the variables that are time lapse dependent come from this class. There are some time dependent blocked log elements in the wells, contained in modelavostatic.cpp. These elements are reconstructed for each time lapse, and they are deleted after each time lapse run.

Some constrains on what is legal for time lapse data are made. The constrains are mostly made when variables are used in modelavostatic.cpp and modelgeneral.cpp, which are called prior to the time lapse loop in main.cpp. The constrains are checked in XmlModelFile::checkTimeLapseConsistency(). 

git-svn-id: https://code.nr.no/scm/svn/crava/branches/StatoilDevelopment@1197 176de4d3-4747-0410-aa96-47f6d6634214