CRA-237, CRA-233, CRA-232 Major changes in CRAVA for time lapse AVO inversion.
main.cpp is restructured to do time lapse inversion.

doinversion.cpp containing setupStaticModels, doFirstAVOInversion and doTimeLapseAVOInversion is added to the project. 

In corr.cpp, a new constructor for time lapse runs is added. In the new constructor, the covariance grids are updated from seismicparametersholder.cpp. In the first run, the covariances are made by multiplying pointwise variances with common correlation grids. In the following runs, the correlations are obtained directly from covariance grids. The deletion of the covariance grids is removed from the destructor, and the grids are deleted in seismicparametersholder.cpp. corr::computeCircCorrT and corr::makeCircCorrTPosDef are moved from FFTGrid. Corr::extractParamCorr, corr::getNextErrorVariance, corr::getNextParameterCovariance, corr::initializeCorrelations, corr::initializeAccess, corr::terminateAccess, corr::createPriorVar0, corr::extractParamCorr, common_correlation_ and errCorr_ are added. 

In crava.cpp, much of what was previously in main.cpp is added. The FFTGrids parSpatialCorr and errCorrUnsmooth are removed from crava.cpp. 

In fftgrid.cpp, the function FFTGrid::fillInErrCorrFromCovAlpha is added. 

A new time lapse constructor is added in ModelAVODynamic. 

seismicparametersholder.cpp is added, containing functions for holding, getting and deleting the expectation and covariance grids. 


git-svn-id: https://code.nr.no/scm/svn/crava/branches/StatoilDevelopment@1210 176de4d3-4747-0410-aa96-47f6d6634214