[CRA-228]
Reimplementation of the time evolution function to be carried out for both static and dynamic parts.

Reimplementation of estimation of transition matrix, A_k. Only the dynamic and static part of the matrix, the interaction between static
and dynamic parts is not implemented. Implemenation of estimation of correction terms, \delta_k and \delta \mu_k.

Combined set up of transition matrix and correction terms in one function, SetUpTimeEvolutionMatrices. In this function a comment on the names
of function parameters is included for - hopefully - easier understanding of the code.
Now all time development matrices/vectors are of dimension 6.

Dummy implementation of split and merge functions for the seismic parameters holder. Possibly misunderstanding of what is wanted.

Change of class variable name for the matrix and vector for the correction term.

git-svn-id: https://code.nr.no/scm/svn/crava/branches/StatoilDevelopment@1263 176de4d3-4747-0410-aa96-47f6d6634214