CRA-341
Reduce memory needed for the multizone background model.

Reduced resolution in the zones, by multiplying dz from the simbox by four. The final multizone background has full resolution.

Make eroded_zones in the for-loop over the zones, instead of making it as a vector outside the loop.

Included logging when building the background model.

git-svn-id: https://code.nr.no/scm/svn/crava/trunk@1392 176de4d3-4747-0410-aa96-47f6d6634214