[CRA-333] Use of rock physics in 3D inversion. In ModelSetting added a boolean variable backgroundFromRockPhysics_ to indicate whether or not background is to be generated from rock physics. Belonging get-function and default initialization to false are also added.
New call to modelGeneral->generateRockPhysics3DBackground() is added in ModelAVODynamic::processBackground(). New output parameter from processBackground() is added to be able to return coefficients to be multiplied with the variance in the correlations structure.

Likewise, in ModelGeneral::processPriorCorrelations, an input parameter is added to handle the coefficients to be multiplied with the variance. 

All new code is deactivated by testing on 0 in if-tests.

NB! The code is not tested with meaningful content.

git-svn-id: https://code.nr.no/scm/svn/crava/branches/StatoilDevelopment@1419 176de4d3-4747-0410-aa96-47f6d6634214