CRA-321 Make sure the order of facies is correct when using rock physics prior models.
Use facies names in iteration over RockDistributions when the pdf's are calculated in FaciesProb::calculateFaciesProbFromRockPhysicsModel() to make sure the facies probabilities are calculated in correct order.

Moved facies_names_ and facies_labels_ and corresponding functions from modelsettings.h to modelgeneral.h. 

Made ModelGeneral::setFaciesNamesFromRockPhysics() and ModelGeneral::setFaciesNamesFromWells(). If facies names are not given in wells, they are taken from the rocks in the rock physics model.

Renamed modelGeneral::processFaciesInformation() to modelGeneral::checkFaciesNamesConsistency(). 

Removed old xml-reading of Gaussian and bounding model that are not to be used.

git-svn-id: https://code.nr.no/scm/svn/crava/branches/StatoilDevelopment@1424 176de4d3-4747-0410-aa96-47f6d6634214