CRA-353 Include HasDistribution() and HasTrend() in DistributionsRock, DistributionsFluid, DistributionsSolid and DistributionsDryRock.
HasDistribution is true if at least one of the variables in the object uses a distribution. At least one of the elements in HasTrend is true if at least one of the variables in the object uses a trend. The element that is true corresponds to the trend cube that is used, where the order of the trend cubes is the same as the order of the cubes in the model file. 

Made class variables is_distribution_ and use_trend_cube_ with corresponding get-functions in distributionwithtrend.h. The former is true if distribution_ is other distribution than dirac. The latter is a vector of bool variables, where the first element is true if the first trend cube is used, the second element is true if the second is used, and both are true if both trend cubes are used. 

Use mostly dummy functions for now. 

git-svn-id: https://code.nr.no/scm/svn/crava/branches/StatoilDevelopment@1429 176de4d3-4747-0410-aa96-47f6d6634214