CRA-414
Make sure all rocks obtain the same number of vintages as the number of seismic vintages

Use n_vintages in GenerateDistributionsFluid(), GenerateDistributionsSolid() and GenerateDistributionsRock() such that the number of vintages is the same for all DistributionsFluid, DistributionsSolid and DistributionsRock. 

When a reservoir variable is evolved, the vintage years need to be given in ascending order, and the years must correspond to the vintage years of the seismic survey. All vintage years must be given whenever a reservoir variable is being evolved.

git-svn-id: https://code.nr.no/scm/svn/crava/branches/StatoilDevelopment@1514 176de4d3-4747-0410-aa96-47f6d6634214